% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNRS.R
\name{TNRS}
\alias{TNRS}
\title{Resolve (plant) taxonomic names}
\usage{
TNRS(taxonomic_names, sources = "tpl,gcc,ildis,tropicos,usda",
  classification = "tropicos", mode = "resolve")
}
\arguments{
\item{taxonomic_names}{}

\item{sources}{Character. Taxonomic sources to use. Default is "tpl,gcc,ildis,tropicos,usda". Options include tpl,ildis,gcc,tropicos,usda,ncbi}

\item{classification}{Character. Family classification to use. Options are tropicos and ncbi. Default is "tropicos", which is equivalent to APGIII.}

\item{mode}{Character.  Options are "resolve" and "parse". Default option is "resolve"}
}
\value{
Dataframe containing TNRS results.
}
\description{
TNRS taxonomically resolved plant taxonomic names.
}
\examples{
\dontrun{
fulldata <- 
read.csv("http://bien.nceas.ucsb.edu/bien/wp-content/uploads/2019/07/tnrs_testfile.csv", header=FALSE)
fulldata$V2 <-gsub(pattern = "'",replacement = "",x = fulldata$V2) #Currently we have an issue with cultivars, so this is a workaround

results <- TNRS(taxonomic_names = fulldata)
  
# Inspect the results
head(results, 10)
}

}
